# swagger.yaml
openapi: 3.0.3
info:
  title: Del-Pick API
  description: API for Del-Pick, a food delivery service application enabling customers to order food, drivers to deliver orders, and store owners to manage their menus and orders.
  version: 1.0.0
  contact:
    name: API Support
    email: support@delpick.id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://delpick.horas-code.my.id/api/v1
    description: Production Server
  - url: http://localhost:6100/api/v1
    description: Development Server
...
paths:
...
  /stores:
    get:
      tags:
        - Stores
      summary: Retrieve all stores
      description: Retrieves a paginated list of all stores, with optional search by name.
      operationId: getAllStores
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/searchParam'
      responses:
        '200':
          description: Stores retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Stores
      summary: Create a new store
      description: Creates a new store and associated owner account. Requires admin privileges.
      operationId: createStore
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreRequest'
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                message: Store created successfully
                data:
                  store:
                    id: 1
                    name: Tasty Bites
                    address: 123 Food Street
                    status: active
                  owner:
                    id: 1
                    name: Store Owner
                    email: owner@example.com
                    role: owner
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stores/{storeId}:
    get:
      tags:
        - Stores
      summary: Retrieve store by ID
      description: Retrieves detailed information about a specific store, including its owner and menu items.
      operationId: getStoreById
      parameters:
        - $ref: '#/components/parameters/storeIdParam'
      responses:
        '200':
          description: Store details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Store details retrieved successfully
                data:
                  store:
                    id: 1
                    name: Tasty Bites
                    address: 123 Food Street
                    status: active
                  owner:
                    id: 1
                    name: Store Owner
                    email: owner@example.com
                  menuItems:
                    - id: 1
                      name: Pizza
                      price: 10.99
                  rating: 4.5
                  reviewCount: 100
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Stores
      summary: Update store
      description: Updates store information. Requires store owner or admin privileges.
      operationId: updateStore
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/storeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreRequest'
      responses:
        '200':
          description: Store updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Store updated successfully
                data:
                  id: 1
                  name: Tasty Bites Updated
                  address: 456 Food Street
                  status: active
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the store owner or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Stores
      summary: Delete store
      description: Deletes a store and associated data. Requires admin or store owner privileges.
      operationId: deleteStore
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/storeIdParam'
      responses:
        '200':
          description: Store deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Store deleted successfully
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the store owner or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /stores/{storeId}/status:
    patch:
      tags:
        - Stores
      summary: Update store status
      description: Updates the status of a store to active or inactive. Requires admin privileges.
      operationId: updateStoreStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/storeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Store status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Store status updated successfully
                data:
                  id: 1
                  name: Tasty Bites
                  status: active
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Requires admin privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Store not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Customer Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    searchParam:
      name: search
      in: query
      description: Search keyword for filtering by name or email
      required: false
      schema:
        type: string
    storeIdParam:
      name: storeId
      in: path
      description: ID of the store
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
      required:
        - statusCode
        - message
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: An error occurred
        errors:
          type: string
          nullable: true
          example: Detailed error message
      required:
        - statusCode
        - message

    # Authentication Schemas
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          example: "08123456789"
        role:
          type: string
          enum: [admin, customer, driver, owner]
          example: customer
        avatar:
          type: string
          format: uri
          example: https://example.com/avatars/john.jpg
          nullable: true
        isVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"

    # Store Schemas
    Store:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
        name:
          type: string
          example: Tasty Bites
        address:
          type: string
          example: 123 Food Street
        description:
          type: string
          example: Best food in town
          nullable: true
        openTime:
          type: string
          example: "09:00"
          nullable: true
        closeTime:
          type: string
          example: "21:00"
          nullable: true
        imageUrl:
          type: string
          format: uri
          example: https://example.com/stores/tasty-bites.jpg
          nullable: true
        phone:
          type: string
          example: "08123456789"
          nullable: true
        latitude:
          type: number
          format: float
          example: -6.12345678
          nullable: true
        longitude:
          type: number
          format: float
          example: 106.12345678
          nullable: true
        distance:
          type: number
          format: float
          example: 2.5
          nullable: true
        status:
          type: string
          enum: [active, inactive]
          example: active
        user:
          $ref: '#/components/schemas/User'
          nullable: true
    StoreListResponse:
      type: object
      properties:
        totalItems:
          type: integer
          example: 50
        totalPages:
          type: integer
          example: 5
        currentPage:
          type: integer
          example: 1
        stores:
          type: array
          items:
            $ref: '#/components/schemas/Store'
...
    CreateStoreRequest:
      type: object
      required:
        - name
        - email
        - password
        - phone
        - storeName
        - address
      properties:
        name:
          type: string
          example: Store Owner
        email:
          type: string
          format: email
          example: owner@example.com
        password:
          type: string
          minLength: 6
          example: password123
        phone:
          type: string
          pattern: '^[0-9]{10,15}$'
          example: "08123456789"
        storeName:
          type: string
          example: Tasty Bites
        address:
          type: string
          example: 123 Food Street
        description:
          type: string
          example: Best food in town
          nullable: true
        openTime:
          type: string
          example: "09:00"
          nullable: true
        closeTime:
          type: string
          example: "21:00"
          nullable: true
        image:
          type: string
          description: Base64 encoded image
          example: data:image/jpeg;base64,/9j/4AAQSkZJRgABAQ...
          nullable: true
        latitude:
          type: number
          format: float
          example: -6.12345678
          nullable: true
        longitude:
          type: number
          format: float
          example: 106.12345678
          nullable: true
    UpdateStoreRequest:
      type: object
      properties:
        name:
          type: string
          example: Store Owner Updated
          nullable: true
        email:
          type: string
          format: email
          example: owner.updated@example.com
          nullable: true
        phone:
          type: string
          pattern: '^[0-9]{10,15}$'
          example: "08987654321"
          nullable: true
        password:
          type: string
          minLength: 6
          example: newpassword123
          nullable: true
        storeName:
          type: string
          example: Tasty Bites Updated
          nullable: true
        address:
          type: string
          example: 456 Food Street
          nullable: true
        description:
          type: string
          example: Updated description
          nullable: true
        openTime:
          type: string
          example: "08:00"
          nullable: true
        closeTime:
          type: string
          example: "22:00"
          nullable: true
        image:
          type: string
          description: Base64 encoded image
          example: data:image/jpeg;base64,/9j/4AAQSkZJRgABAQ...
          nullable: true
        latitude:
          type: number
          format: float
          example: -6.12445678
          nullable: true
        longitude:
          type: number
          format: float
          example: 106.12445678
          nullable: true
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [active, inactive]
          example: active

    # Customer Schemas
...